Fix bug in gralloc
If buffers were allocated and then locked without ever
being passed to another process lock would return NULL
instead of a pointer to the buffer.  Instead map the buffer
from lock if it hasn't already been done by register.  Also
check in free and unmap the buffer it is mapped.

Change-Id: I3fe670ccb429f3de5f2e4dd07768414e70a02036
Signed-off-by: Rebecca Schultz Zavin <rebecca@android.com>